requirements/dev.txt
-r base.txt

# Dev tools
django-debug-toolbar>=4.0,<5.0
django-extensions>=3.2,<4.0

ipython>=8.0,<9.0
black>=24.0,<25.0
isort>=5.13,<6.0
flake8>=7.0,<8.0

pytest>=8.0,<9.0
pytest-django>=4.8,<5.0